+++ example/lib/example5/widgets/panels/app_bar_settings/app_bar_style_buttons.dart
@@ -2,6 +2,10 @@ import 'package:flex_color_scheme/flex_color_scheme.dart';
 import 'package:flutter/material.dart';
 
 // Widget used to change the used app bar style mode in example 5.
+//
+// In version 5.2.0 of FlexColorScheme and its example 5, this widget is
+// no longer used, but keeping it around for a while.
+// TODO(rydmike): Remove this widget in example version 6.0.0.
 class AppBarStyleButtons extends StatelessWidget {
   const AppBarStyleButtons({
     super.key,